requirement.txt
django
djangorestframework
psycopg2
django-cors-headers
djangorestframework-simplejwt
bigcommerce
django-environ
stripe
ShopifyAPI
django-crontab